+++ es6/rules.js
@@ -3,14 +3,15 @@ module.exports = {
     // Allowing groups
     'sort-imports': ['off', { allowSeparatedGroups: true }],
 
-    // false positive prone, https://github.com/nodesecurity/eslint-plugin-security/issues/21
-    'security/detect-object-injection': 'off',
-
     // Allow unused args with underscore
     'no-unused-vars': ['error', { argsIgnorePattern: '^_' }],
 
     // conflicts with prettier
     'array-bracket-spacing': 'off',
-    'max-len': 'off'
+    'max-len': 'off',
+
+    // overly strict security rules
+    'security/detect-object-injection': 'off', // false positive prone, https://github.com/nodesecurity/eslint-plugin-security/issues/21
+    'security/detect-non-literal-fs-filename': 'off' // many false positives
   }
 };