+++ CHANGELOG.md
@@ -6,5 +6,6 @@
 - Minor: Allow unused args prefixed with underscore
 - Minor: Fix `@typescript-eslint/no-shadow`
 - Minor: Fix `@typescript-eslint/no-non-null-assertion`
+- Minor: Disable `security/detect-non-literal-fs-filename`
 
 ---